+++ src/export_roms.rs
@@ -1,6 +1,6 @@
 use super::bchunk;
 use super::chdman;
-use super::chdman::{AsChd, ToChd};
+use super::chdman::{AsChd, MediaType, ToChd};
 use super::common::*;
 use super::config::*;
 use super::database::*;
@@ -810,7 +810,7 @@ async fn to_chd(
                 continue;
             }
         }
-        let (cue_roms, bin_roms): (Vec<&Rom>, Vec<&Rom>) = roms
+        let (cue_roms, bin_iso_roms): (Vec<&Rom>, Vec<&Rom>) = roms
             .into_par_iter()
             .partition(|rom| rom.name.ends_with(CUE_EXTENSION));
 
@@ -823,9 +823,9 @@ async fn to_chd(
                     .await?,
             );
         }
-        let mut bin_romfiles: Vec<CommonRomfile> = Vec::new();
-        for rom in &bin_roms {
-            bin_romfiles.push(
+        let mut bin_iso_romfiles: Vec<CommonRomfile> = Vec::new();
+        for rom in &bin_iso_roms {
+            bin_iso_romfiles.push(
                 romfile
                     .as_archive(rom)?
                     .to_common(progress_bar, &tmp_directory.path())
@@ -836,20 +836,25 @@ async fn to_chd(
             Some(cue_romfile) => {
                 cue_romfile
                     .as_cue_bin(
-                        &bin_romfiles
+                        &bin_iso_romfiles
                             .iter()
                             .map(|bin_iso_romfile| &bin_iso_romfile.path)
                             .collect::<Vec<&PathBuf>>(),
                     )?
-                    .to_chd(progress_bar, destination_directory, &Some(cue_romfile))
+                    .to_chd(
+                        progress_bar,
+                        destination_directory,
+                        &Some(cue_romfile),
+                        &MediaType::Cd,
+                    )
                     .await?
             }
             None => {
-                bin_romfiles
+                bin_iso_romfiles
                     .first()
                     .unwrap()
                     .as_iso()?
-                    .to_chd(progress_bar, destination_directory, &None)
+                    .to_chd(progress_bar, destination_directory, &None, &MediaType::Dvd)
                     .await?
             }
         };
@@ -878,6 +883,7 @@ async fn to_chd(
                 progress_bar,
                 destination_directory,
                 &Some(&cue_romfile.as_common()?),
+                &MediaType::Cd,
             )
             .await?;
     }
@@ -888,7 +894,7 @@ async fn to_chd(
             let romfile = romfiles_by_id.get(&rom.romfile_id.unwrap()).unwrap();
             romfile
                 .as_iso()?
-                .to_chd(progress_bar, destination_directory, &None)
+                .to_chd(progress_bar, destination_directory, &None, &MediaType::Dvd)
                 .await?;
         }
     }
@@ -902,7 +908,7 @@ async fn to_chd(
                 .as_xso()?
                 .to_iso(progress_bar, &tmp_directory.path())
                 .await?
-                .to_chd(progress_bar, destination_directory, &None)
+                .to_chd(progress_bar, destination_directory, &None, &MediaType::Dvd)
                 .await?;
         }
     }
@@ -916,7 +922,7 @@ async fn to_chd(
                 .as_xso()?
                 .to_iso(progress_bar, &tmp_directory.path())
                 .await?
-                .to_chd(progress_bar, destination_directory, &None)
+                .to_chd(progress_bar, destination_directory, &None, &MediaType::Dvd)
                 .await?;
         }
     }